fix: emit `-std=c++17` to compile_commands.json.
We are using gcc 13. By default, it uses `-std=gnu++17`.
So, in cmake, if we just write:

    set(CMAKE_CXX_STANDARD 17)

cmake won't emit any `-std=...` flag, thus the flag is missing in compile_commands.json.

However, for auto-completion with ccls to work, it requires an explicit `-std=...` flag.
This commit makes cmake to emit the required flag.